requirements.txt
# PDF Processing
markitdown>=0.1.0
PyMuPDF>=1.23.0
pdf2image>=1.16.0

# LLM Clients
google-generativeai>=0.3.0
openai>=1.0.0
anthropic>=0.18.0

# Utilities
pyyaml>=6.0
python-dotenv>=1.0.0

# OCR (optional)
pytesseract>=0.3.10
google-cloud-vision>=3.4.0

# Async & Parallel Processing
aiohttp>=3.9.0
tenacity>=8.2.0
